+++ modules/apps/portal-security/portal-security-antisamy/src/main/java/com/liferay/portal/security/antisamy/internal/AntiSamySanitizerPublisher.java
@@ -16,8 +16,8 @@ package com.liferay.portal.security.antisamy.internal;
 
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.sanitizer.Sanitizer;
-import com.liferay.portal.security.antisamy.configuration.AntiSamyClassNameConfiguration;
-import com.liferay.portal.security.antisamy.configuration.AntiSamyConfiguration;
+import com.liferay.portal.security.antisamy.internal.configuration.AntiSamyClassNameConfiguration;
+import com.liferay.portal.security.antisamy.internal.configuration.AntiSamyConfiguration;
 
 import java.net.URL;
